migrated plugin code from grails 1.3.7 to 3.0.2 except test cases.
