--- application.properties
@@ -1,4 +1,0 @@
-#Grails Metadata file
-#Sat Jun 05 12:27:39 PDT 2010
-app.grails.version=1.3.6
-app.name=csv
