+++ gradle.properties
@@ -1,0 +1,2 @@
+grailsVersion=3.0.2
+gradleWrapperVersion=2.3
