--- plugin.xml
@@ -1,114 +1,0 @@
-<plugin name='csv' version='0.3' grailsVersion='1.2 &gt; *'>
-  <author>Les Hazlewood</author>
-  <authorEmail>les@katasoft.com</authorEmail>
-  <title>Grails CSV Plugin</title>
-  <description>
-        The Grails CSV Plugin allows you to easily parse and consume CSV data from a number of input sources.  It
-        supports complex parsing scenarios such as nested commas inside quotes, escaped tokens, multi-line quoted
-        values and allows configuration of parsing options (separator char, escape char, text encoding, etc).  It is
-        based on Glen Smith (et. al.)'s OpenCSV project (http://opencsv.sourceforge.net)
-
-        This plugin adds two dynamic methods 'eachCsvLine' and 'toCsvReader' to each of the following classes:
-        - java.lang.String
-        - java.io.File
-        - java.io.InputStream
-        - java.io.Reader
-
-        Using it is extremely simple.  On any instance of the four data types, call the 'eachCvsLine' method with a
-          closure accepting the tokens (a String array) for each parsed line:
-
-        "hello, world, how, are, you".eachCsvLine { tokens -&gt;
-            //only one line in this case and tokens.length == 5
-        }
-
-        new File("iso3166Countries.csv").eachCsvLine { tokens -&gt;
-            new Country(tokens[0],        //ISO 3166 country name
-                        tokens[1]).save() //ISO 3166 2 letter character code
-        }
-
-        Configuration
-
-        If you need to specify how the parsing should occur, you can construct your own csv reader with a map of
-        configuration options and call the 'eachLine' method on the constructed reader:
-
-        anInputStream.toCsvReader(['charset':'UTF-8']).eachLine { tokens -&gt;
-            ...
-        }
-
-        The supported config options:
-
-        'separatorChar': the character to use as the delimiter to separate the tokens.  Defaults to the comma: ','
-
-        'quoteChar': the character indicating a quoted string is about to follow.  Internal separatorChars can be
-                     inside the quoted string and they will not be split into tokens.
-                     Defaults to the double quote char: '"'
-
-        'escapeChar': the character to escape an immediately following character, indicating to the parser not to treat
-                      it as a special char.  Defaults to the backslash char: ''
-
-        'skipLines': the number of lines in the input source to skip before parsing begins.  This is useful to skip
-                     any potential CSV header lines that are not part of the CSV data.  Defaults to zero '0'
-
-        'strictQuotes': if characters outside of quotes should be ignored (implying each individual token is
-                        quoted.  Defaults to false
-
-        'ignoreLeadingWhiteSpace': white space in front of a quoted token is ignored.  Defaults to true
-
-        'charset': use the specified charset when parsing an InputStream.  The value can be either the Charset name
-                   as a String, a java.nio.charset.Charset instance, or a java.nio.charset.CharsetDecoder instance.
-                   Defaults to the system default charset.
-                   *Note that this option is ONLY valid for InputStream instances.  It is ignored otherwise.
-
-    </description>
-  <documentation>http://grails.org/plugin/csv</documentation>
-  <resources>
-    <resource>BuildConfig</resource>
-    <resource>DataSource</resource>
-    <resource>UrlMappings</resource>
-    <resource>CsvTestController</resource>
-  </resources>
-  <dependencies />
-  <behavior>
-    <method name='eachLine' artefact='Controller' type='au.com.bytecode.opencsv.CSVReader'>
-      <description />
-      <argument type='java.lang.Object' />
-    </method>
-    <method name='eachCsvLine' artefact='Controller' type='java.io.InputStream'>
-      <description />
-      <argument type='java.lang.Object' />
-    </method>
-    <method name='toCsvReader' artefact='Controller' type='java.io.InputStream'>
-      <description />
-      <argument type='java.lang.Object' />
-    </method>
-    <method name='eachCsvLine' artefact='Controller' type='java.io.Reader'>
-      <description />
-      <argument type='java.lang.Object' />
-    </method>
-    <method name='toCsvReader' artefact='Controller' type='java.io.Reader'>
-      <description />
-      <argument type='java.lang.Object' />
-    </method>
-    <method name='toCsvMapReader' artefact='Controller' type='java.io.Reader'>
-      <description />
-      <argument type='java.lang.Object' />
-    </method>
-    <method name='eachCsvLine' artefact='Controller' type='java.lang.String'>
-      <description />
-      <argument type='java.lang.Object' />
-    </method>
-    <method name='toCsvReader' artefact='Controller' type='java.lang.String'>
-      <description />
-      <argument type='java.lang.Object' />
-    </method>
-    <method name='toCsvMapReader' artefact='Controller' type='java.lang.String'>
-      <description />
-      <argument type='java.lang.Object' />
-    </method>
-    <method name='renderCsv' artefact='Controller' type='CsvTestController'>
-      <description />
-      <argument type='java.util.Map' />
-      <argument type='groovy.lang.Closure' />
-    </method>
-  </behavior>
-</plugin>
