+++ src/test/groovy/grails/csv/CsvTestControllerSpec.groovy
@@ -1,0 +1,22 @@
+package grails.csv
+
+import grails.test.mixin.TestFor
+import spock.lang.Specification
+
+/**
+ * See the API for {@link grails.test.mixin.web.ControllerUnitTestMixin} for usage instructions
+ */
+@TestFor(CsvTestController)
+class CsvTestControllerSpec extends Specification {
+
+    def setup() {
+    }
+
+    def cleanup() {
+    }
+
+    void "test something"() {
+        expect:"fix me"
+            true == false
+    }
+}
